+++ src/andyroo/agility/Obstacle.java
@@ -1,4 +1,4 @@
-package andyroo;
+package andyroo.agility;
 
 
 public class Obstacle {
